Reorder fields by inserting field into fields dict
Django 1.7 has changed the underlying data structure of form.fields from its internal SortedDict to the standard lib OrderedDict. OrderedDict does not have the keyOrder attribute we were using previously to move the 'email' field on the login form in front of the 'password' field (even though it was inserted afterwards).
In the case of an OrderedDict, we now just pop the 'password' field and reinsert it at the end to reorder them.